+++ Snippetu/Sound2Light/Snippetu_fastLedAudioIn/fragmentColumnMovingRight.h
@@ -13,7 +13,7 @@ class ColumnMovingRight : public Fragment
     public:
         ColumnMovingRight(LedUtils ledUtils, unsigned long time);
         void setAnimationProperties(int length=0, int start=0, int wait=10, int fade=250);
-        void setColorProperties(double hue=0, int sat=0, int bri=255, double hueIncrement=0);
+        void setColorProperties(uint8_t hue=0, uint8_t sat=0, uint8_t bri=255, double hueIncrement=0);
         void nextFrame(unsigned long currentTime, std::vector<PixelUpdate> &matrixUpdate, bool &animationFinished);
         void resetTimer(unsigned long time);
         String getValues();
